+++ CHANGELOG.md
@@ -78,6 +78,11 @@ All notable changes to this project will be documented in this file.
 
 # Develop
 
+## Android
+
+### Refactor
+* Rename BookmarkListScreenlet comparator to obcClassName
+
 ## iOS
 
 ### Bugs